# mdijk.blazoring/Shared/Models/Driver.cs
namespace mdijk.blazoring.Shared.Models
{
	public class Driver
	{		
		public int DriverId { get; set; }
		public string FirstName { get; set; }
		public string LastName { get; set; }
		public string Country { get; set; }
		public int DriverNumber { get; set; }
	}
}