# mdijk.blazoring/swagger.json
{
...
  "swagger": "2.0",
  "info": {
    "title": "My Title",
    "version": "1.0.0"
  },
  "host": "localhost:57256",
  "schemes": [
    "http"
  ],
  "produces": [
    "text/plain",
    "application/json",
    "text/json"
  ],
  "paths": {
...
    "/Season/addSeason": {
      "post": {
        "tags": [
          "Season"
        ],
        "operationId": "Season_AddSeason",
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "name": "Season",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Season"
            },
            "x-nullable": false
          }
        ],
        "responses": {
          "200": {
            "x-nullable": false,
            "description": "",
            "schema": {
              "$ref": "#/definitions/AddItemResponse"
            }
          }
        }
      }
    },
    "/Season/register": {
      "post": {
        "tags": [
          "Season"
        ],
        "operationId": "Season_Register",
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "name": "model",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/RegisterModel"
            },
            "x-nullable": false
          }
        ],
        "responses": {
          "200": {
            "x-nullable": false,
            "description": "",
            "schema": {
              "$ref": "#/definitions/AddItemResponse"
            }
          }
        }
      }
    },
    "/Season/addentry": {
      "post": {
        "tags": [
          "Season"
        ],
        "operationId": "Season_AddEntry",
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "name": "entryModel",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/AddEntryModel"
            },
            "x-nullable": false
          }
        ],
        "responses": {
          "200": {
            "x-nullable": false,
            "description": "",
            "schema": {
              "$ref": "#/definitions/AddItemResponse"
            }
          }
        }
      }
    },
    "/Season/getseason/{seasonId}/{seasonYear}": {
      "get": {
        "tags": [
          "Season"
        ],
        "operationId": "Season_GetSeason",
        "parameters": [
          {
            "type": "integer",
            "name": "seasonId",
            "in": "path",
            "required": true,
            "format": "int32",
            "x-nullable": false
          },
          {
            "type": "integer",
            "name": "seasonYear",
            "in": "path",
            "required": true,
            "format": "int32",
            "x-nullable": false
          }
        ],
        "responses": {
          "200": {
            "x-nullable": false,
            "description": "",
            "schema": {
              "$ref": "#/definitions/Season"
            }
          }
        }
      }
    },
    "/Season/getseasoncalendar/{seasonId}/{seasonYear}": {
      "get": {
        "tags": [
          "Season"
        ],
        "operationId": "Season_GetSeasonCalendar",
        "parameters": [
          {
            "type": "integer",
            "name": "seasonId",
            "in": "path",
            "required": true,
            "format": "int32",
            "x-nullable": false
          },
          {
            "type": "integer",
            "name": "seasonYear",
            "in": "path",
            "required": true,
            "format": "int32",
            "x-nullable": false
          }
        ],
        "responses": {
          "200": {
            "x-nullable": false,
            "description": "",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/RaceEvent"
              }
            }
          }
        }
      }
    },
    "/Season/addraceevent": {
      "post": {
        "tags": [
          "Season"
        ],
        "operationId": "Season_AddRaceEvent",
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "name": "raceEventModel",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/AddRaceEventModel"
            },
            "x-nullable": false
          }
        ],
        "responses": {
          "200": {
            "x-nullable": false,
            "description": "",
            "schema": {
              "$ref": "#/definitions/AddItemResponse"
            }
          }
        }
      }
    },
    "/Season/getentriesforevent/{eventId}": {
      "get": {
        "tags": [
          "Season"
        ],
        "operationId": "Season_GetEntriesForEvent",
        "parameters": [
          {
            "type": "integer",
            "name": "eventId",
            "in": "path",
            "required": true,
            "format": "int32",
            "x-nullable": false
          }
        ],
        "responses": {
          "200": {
            "x-nullable": false,
            "description": "",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Entry"
              }
            }
          }
        }
      }
    },
...
    "/Season/seasonresults/{seasonId}": {
      "get": {
        "tags": [
          "Season"
        ],
        "operationId": "Season_GetSeasonResults",
        "parameters": [
          {
            "type": "integer",
            "name": "seasonId",
            "in": "path",
            "required": true,
            "format": "int32",
            "x-nullable": false
          }
        ],
        "responses": {
          "200": {
            "x-nullable": false,
            "description": "",
            "schema": {
              "$ref": "#/definitions/SeasonResultsModel"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
    "Circuit": {
      "type": "object",
      "required": [
        "circuitId"
      ],
      "properties": {
        "circuitId": {
          "type": "integer",
          "format": "int32"
        },
        "circuitName": {
          "type": "string"
        },
        "country": {
          "type": "string"
        }
      }
    },
    "AddItemResponse": {
      "allOf": [
        {
          "$ref": "#/definitions/ResponseModel"
        },
        {
          "type": "object",
          "required": [
            "newItemId"
          ],
          "properties": {
            "newItemId": {
              "type": "integer",
              "format": "int32"
            }
          }
        }
      ]
    },
    "ResponseModel": {
      "type": "object",
      "required": [
        "isError"
      ],
      "properties": {
        "message": {
          "type": "string"
        },
        "isError": {
          "type": "boolean"
        }
      }
    },
...
    "Driver": {
      "type": "object",
      "required": [
        "driverId",
        "driverNumber"
      ],
      "properties": {
        "driverId": {
          "type": "integer",
          "format": "int32"
        },
        "firstName": {
          "type": "string"
        },
        "lastName": {
          "type": "string"
        },
        "country": {
          "type": "string"
        },
        "driverNumber": {
          "type": "integer",
          "format": "int32"
        }
      }
    },
...
    "Engine": {
      "type": "object",
      "required": [
        "engineId"
      ],
      "properties": {
        "engineId": {
          "type": "integer",
          "format": "int32"
        },
        "engineName": {
          "type": "string"
        },
        "country": {
          "type": "string"
        }
      }
    },
    "EventResult": {
      "type": "object",
      "required": [
        "eventResultId",
        "finishingPosition",
        "fastestLap"
      ],
      "properties": {
        "eventResultId": {
          "type": "integer",
          "format": "int32"
        },
        "finishingPosition": {
          "type": "integer",
          "format": "int32"
        },
        "fastestLap": {
          "type": "boolean"
        },
        "driverEntry": {
          "$ref": "#/definitions/DriverEntry"
        },
        "metaData": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/EventResultMetaData"
          }
        },
        "typedMetaData": {
          "$ref": "#/definitions/EventResultMetaDataTyped"
        }
      }
    },
    "DriverEntry": {
      "type": "object",
      "required": [
        "driverEntryId"
      ],
      "properties": {
        "driverEntryId": {
          "type": "integer",
          "format": "int32"
        },
        "driver": {
          "$ref": "#/definitions/Driver"
        },
        "entryFullEntryName": {
          "type": "string"
        },
        "entryTeam": {
          "$ref": "#/definitions/Team"
        },
        "entryEngine": {
          "$ref": "#/definitions/Engine"
        }
      }
    },
    "Team": {
      "type": "object",
      "required": [
        "teamId"
      ],
      "properties": {
        "teamId": {
          "type": "integer",
          "format": "int32"
        },
        "teamName": {
          "type": "string"
        },
        "country": {
          "type": "string"
        }
      }
    },
    "EventResultMetaData": {
      "type": "object",
      "required": [
        "eventResultMetaDataId"
      ],
      "properties": {
        "eventResultMetaDataId": {
          "type": "integer",
          "format": "int32"
        },
        "key": {
          "type": "string"
        },
        "value": {
          "type": "string"
        }
      }
    },
    "EventResultMetaDataTyped": {
      "type": "object",
      "required": [
        "gridPosition",
        "laps",
        "fastestLapRank"
      ],
      "properties": {
        "points": {
          "type": "string"
        },
        "gridPosition": {
          "type": "integer",
          "format": "int32"
        },
        "laps": {
          "type": "integer",
          "format": "int32"
        },
        "status": {
          "type": "string"
        },
        "time": {
          "type": "string"
        },
        "fastestLap": {
          "type": "string"
        },
        "fastestLapRank": {
          "type": "integer",
          "format": "int32"
        }
      }
    },
...
    "RaceEvent": {
      "type": "object",
      "required": [
        "raceEventId",
        "sequenceNumber"
      ],
      "properties": {
        "raceEventId": {
          "type": "integer",
          "format": "int32"
        },
        "sequenceNumber": {
          "type": "integer",
          "format": "int32"
        },
        "raceEventName": {
          "type": "string"
        },
        "circuit": {
          "$ref": "#/definitions/Circuit"
        },
        "season": {
          "$ref": "#/definitions/Season"
        },
        "results": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/EventResult"
          }
        }
      }
    },
    "Season": {
      "type": "object",
      "required": [
        "seasonId",
        "seasonYear"
      ],
      "properties": {
        "seasonId": {
          "type": "integer",
          "format": "int32"
        },
        "seasonYear": {
          "type": "integer",
          "format": "int32"
        },
        "entries": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Entry"
          }
        },
        "raceEvents": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/RaceEvent"
          }
        }
      }
    },
    "Entry": {
      "type": "object",
      "required": [
        "entryId"
      ],
      "properties": {
        "entryId": {
          "type": "integer",
          "format": "int32"
        },
        "fullEntryName": {
          "type": "string"
        },
        "team": {
          "$ref": "#/definitions/Team"
        },
        "drivers": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/DriverEntry"
          }
        },
        "engine": {
          "$ref": "#/definitions/Engine"
        }
      }
    },
    "RegisterModel": {
      "type": "object",
      "required": [
        "seasonId",
        "teamId",
        "driverId"
      ],
      "properties": {
        "seasonId": {
          "type": "integer",
          "format": "int32"
        },
        "teamId": {
          "type": "integer",
          "format": "int32"
        },
        "driverId": {
          "type": "integer",
          "format": "int32"
        },
        "entryName": {
          "type": "string"
        }
      }
    },
    "AddEntryModel": {
      "type": "object",
      "required": [
        "entryId",
        "seasonId"
      ],
      "properties": {
        "entryId": {
          "type": "integer",
          "format": "int32"
        },
        "fullEntryName": {
          "type": "string"
        },
        "teamId": {
          "type": "string"
        },
        "firstDriverId": {
          "type": "string"
        },
        "secondDriverId": {
          "type": "string"
        },
        "engineId": {
          "type": "string"
        },
        "seasonId": {
          "type": "integer",
          "format": "int32"
        }
      }
    },
    "AddRaceEventModel": {
      "type": "object",
      "required": [
        "raceEventId",
        "seasonId"
      ],
      "properties": {
        "raceEventId": {
          "type": "integer",
          "format": "int32"
        },
        "sequenceNr": {
          "type": "string"
        },
        "raceEventName": {
          "type": "string"
        },
        "circuitId": {
          "type": "string"
        },
        "seasonId": {
          "type": "integer",
          "format": "int32"
        }
      }
    },
    "SeasonResultsModel": {
      "type": "object",
      "required": [
        "seasonId",
        "seasonYear"
      ],
      "properties": {
        "seasonId": {
          "type": "integer",
          "format": "int32"
        },
        "seasonYear": {
          "type": "integer",
          "format": "int32"
        },
        "raceEvents": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/SeasonRaceEvent"
          }
        }
      }
    },
    "SeasonRaceEvent": {
      "type": "object",
      "required": [
        "raceEventId",
        "sequenceNumber"
      ],
      "properties": {
        "raceEventId": {
          "type": "integer",
          "format": "int32"
        },
        "sequenceNumber": {
          "type": "integer",
          "format": "int32"
        },
        "raceEventName": {
          "type": "string"
        },
        "circuit": {
          "$ref": "#/definitions/Circuit"
        },
        "results": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/SeasonRaceEventResult"
          }
        }
      }
    },
    "SeasonRaceEventResult": {
      "type": "object",
      "required": [
        "eventResultId",
        "finishingPosition",
        "fastestLap"
      ],
      "properties": {
        "eventResultId": {
          "type": "integer",
          "format": "int32"
        },
        "finishingPosition": {
          "type": "integer",
          "format": "int32"
        },
        "fastestLap": {
          "type": "boolean"
        },
        "driverEntry": {
          "$ref": "#/definitions/DriverEntry"
        },
        "metaData": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/EventResultMetaData"
          }
        }
      }
    },
...
  }
}